requirements.txt
torch
torchvision
open_clip_torch
Pillow
scikit-learn
datasets
matplotlib